+++ pkg/exporter/sumologicexporter/config_test.go
@@ -19,7 +19,7 @@ func TestInitExporterInvalidConfiguration(t *testing.T) {
 			expectedError: errors.New("unexpected log format: test_format"),
 			cfg: &Config{
 				LogFormat:        "test_format",
-				MetricFormat:     "carbon2",
+				MetricFormat:     "otlp",
 				CompressEncoding: "gzip",
 				TraceFormat:      "otlp",
 				HTTPClientSettings: confighttp.HTTPClientSettings{
@@ -42,11 +42,24 @@ func TestInitExporterInvalidConfiguration(t *testing.T) {
 			},
 		},
 		{
+			name:          "unsupported Carbon2 metrics format",
+			expectedError: errors.New("support for the carbon2 metric format was removed, please use prometheus or otlp instead"),
+			cfg: &Config{
+				LogFormat:    "json",
+				MetricFormat: "carbon2",
+				HTTPClientSettings: confighttp.HTTPClientSettings{
+					Timeout:  defaultTimeout,
+					Endpoint: "test_endpoint",
+				},
+				CompressEncoding: "gzip",
+			},
+		},
+		{
 			name:          "unexpected trace format",
 			expectedError: errors.New("unexpected trace format: text"),
 			cfg: &Config{
 				LogFormat:    "json",
-				MetricFormat: "carbon2",
+				MetricFormat: "otlp",
 				TraceFormat:  "text",
 				HTTPClientSettings: confighttp.HTTPClientSettings{
 					Timeout:  defaultTimeout,
@@ -60,7 +73,7 @@ func TestInitExporterInvalidConfiguration(t *testing.T) {
 			expectedError: errors.New("invalid compression encoding type: test_format"),
 			cfg: &Config{
 				LogFormat:        "json",
-				MetricFormat:     "carbon2",
+				MetricFormat:     "otlp",
 				CompressEncoding: "test_format",
 				TraceFormat:      "otlp",
 				HTTPClientSettings: confighttp.HTTPClientSettings{
@@ -74,7 +87,7 @@ func TestInitExporterInvalidConfiguration(t *testing.T) {
 			expectedError: errors.New("no endpoint and no auth extension specified"),
 			cfg: &Config{
 				LogFormat:        "json",
-				MetricFormat:     "carbon2",
+				MetricFormat:     "otlp",
 				CompressEncoding: "gzip",
 				TraceFormat:      "otlp",
 				HTTPClientSettings: confighttp.HTTPClientSettings{