+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Breaking changes
+
+- feat(sumologicexporter)!: remove support for Carbon2 metrics format [#590][#590]
+
 [Unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.51.0-sumo-0...main
+[#590]: https://github.com/SumoLogic/sumologic-otel-collector/pull/590
 
 ## [v0.51.0-sumo-0]